+++ packages/Peptides/src/widgets/settings.ts
@@ -19,6 +19,7 @@ export enum SETTINGS_PANES {
 };
 
 export enum GENERAL_INPUTS {
+  ACTIVITY = 'Activity',
   ACTIVITY_SCALING = 'Activity scaling',
   BIDIRECTIONAL_ANALYSIS = 'Bidirectional analysis',
 }
@@ -58,14 +59,17 @@ export function getSettingsDialog(model: PeptidesModel): SettingsElements {
   const inputs: PaneInputs = {};
 
   // General pane options
+  const activityCol = ui.columnInput(GENERAL_INPUTS.ACTIVITY, model.df,
+    model.df.getCol(model.settings.activityColumnName!), () => result.activityColumnName = activityCol.value!.name,
+    {filter: (col: DG.Column) => (col.type === DG.TYPE.FLOAT || col.type === DG.TYPE.INT) && col.name !== C.COLUMNS_NAMES.ACTIVITY_SCALED});
   const activityScaling =
     ui.choiceInput(GENERAL_INPUTS.ACTIVITY_SCALING, currentScaling, Object.values(C.SCALING_METHODS),
       () => result.scaling = activityScaling.value as C.SCALING_METHODS) as DG.InputBase<C.SCALING_METHODS>;
   const bidirectionalAnalysis = ui.boolInput(GENERAL_INPUTS.BIDIRECTIONAL_ANALYSIS, currentBidirectional,
     () => result.isBidirectional = bidirectionalAnalysis.value) as DG.InputBase<boolean>;
 
-  accordion.addPane(SETTINGS_PANES.GENERAL, () => ui.inputs([activityScaling, bidirectionalAnalysis]), true);
-  inputs[SETTINGS_PANES.GENERAL] = [activityScaling, bidirectionalAnalysis];
+  accordion.addPane(SETTINGS_PANES.GENERAL, () => ui.inputs([activityCol, activityScaling, bidirectionalAnalysis]), true);
+  inputs[SETTINGS_PANES.GENERAL] = [activityCol, activityScaling, bidirectionalAnalysis];
 
   // Viewers pane options
   /* FIXME: combinations of adding and deleting viewers are not working properly